do not call Task::notify() ourself!
ref[0]

The call to notify() was added to "fix" a hang in the rmp-rpc[1] crate.
I thought that I had to explicitely tell the executor to re-poll the
Endpoint because at some point the Endpoint was not being polled
anymore. While that fixes the hanging issue, it made the Endpoint being
constantly polled which ate all the CPU.

The thing is, when using the Tokio event loop, everything is handled for
us: the Endpoint is polled only when some IO event allows it to makes
some progress. Calling notify() ourself is defeating the purpose of the
Tokio event loop.

This commit also fixes a bug @Fylwind found while helping me debugging:
we have to read all the messages from the stream before returning
Async::NotReady, otherwise, the Endpoint won't be polled before a new
message arrives, hence leaving these messages unread for an undetermined
period of time.

[0]
https://users.rust-lang.org/t/my-future-eats-all-the-cpu-how-can-i-prevent-this/13084/7

[1] https://github.com/little-dude/rmp-rpc