refactor endpoint
1. Introduce Intostaticfuture

This allows us to remove a box from the Service trait:
handle_request() and handle_notification() now return an
IntoStaticfuture instead of a Box<dyn Future>.

2. Simplify the return type in the Service trait

Service::handle_request() now returns a Future<Item=T, Error=E>
instead of a Future<Item=Result<T, E>, Error=SomeError>. The reason
for this is that there is no reason in the Endpoint to handle
Future::Error cases, because there's nothing we can do about it apart
from logging them. It's better to remove one layer, and use the
Future::Error to carry the actual error the handler wants to send to
its peer.

Similarly, Service::handle_notification() now returns a
Future<Item=(), Error=()>. We really don't need the Future::Error case
here, but that's how Future works in 0.1 so we roll with it (although
this may change with future 0.3).

3. Do not store and poll the Server's futures

The Server use to store the request and notification tasks, and we
would poll them manually, but that's what the Executor is for
actually. It's more idiomatic to spawn those tasks, and keep a channel
open to receive the responses from the requests that the service
finished handling like we now do in Server::process_request. This
probably also avoids some allocations (we got rid of a HashMap and a
Vec).